requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.7.1
dj-rest-auth[with_social]==4.0.0
Django==5.1.3
django-allauth==0.52.0
djangorestframework==3.15.2
idna==3.10
oauthlib==3.2.2
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
drf-yasg==1.21.8
djangorestframework-simplejwt==5.3.1
django-filter==24.3
config~=0.5.1
celery~=5.3.1
redis
flower
drf-writable-nested==0.7.1
boto3==1.35.77
django-storages==1.14.4
Pillow==11.0.0
